# docs/swagger.yaml
basePath: /
definitions:
...
  models.ApiError:
    properties:
      error:
        type: string
    type: object
  models.Genre:
    properties:
      id:
        type: integer
      title:
        type: string
    type: object
  models.Movie:
    properties:
      description:
        type: string
      director:
        type: string
      genres:
        items:
          $ref: '#/definitions/models.Genre'
        type: array
      id:
        type: integer
      isWatched:
        type: boolean
      posterUrl:
        type: string
      rating:
        type: integer
      releaseYear:
        type: integer
      title:
        type: string
      trailerUrl:
        type: string
    type: object
host: localhost:8081
info:
  contact:
    email: bexultanovmadiyar@gmail.com
    name: Madiyar Bexultanov
    url: https://www.linkedin.com/in/madiyar-bexultanov-b21902258/
  description: Personal online platform providing information about films
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Ozinshe API
  version: "1.0"
paths:
...
  /movies:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: genreId
        type: string
      - in: query
        name: isWatched
        type: string
      - in: query
        name: searchTerm
        type: string
      - in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Movie'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - Bearer: []
      summary: Get all movies
      tags:
      - movies
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Title
        in: formData
        name: title
        required: true
        type: string
      - description: Description
        in: formData
        name: description
        required: true
        type: string
      - description: Year of release
        in: formData
        name: releaseYear
        required: true
        type: integer
      - description: Director
        in: formData
        name: director
        required: true
        type: string
      - description: Trailer URL
        in: formData
        name: trailerUrl
        required: true
        type: string
      - collectionFormat: csv
        description: Genre ids
        in: formData
        items:
          type: integer
        name: genreIds
        required: true
        type: array
      - description: Poster image
        in: formData
        name: poster
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              id:
                type: integer
            type: object
        "400":
          description: Invalid data
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - Bearer: []
      summary: Create movie
      tags:
      - movies
  /movies/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Movie id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid data
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - Bearer: []
      summary: Delete movie
      tags:
      - movies
    get:
      consumes:
      - application/json
      parameters:
      - description: Movie id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Movie'
        "400":
          description: Invalid movie id
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - Bearer: []
      summary: Find by id
      tags:
      - movies
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: Movie id
        in: path
        name: id
        required: true
        type: integer
      - description: Title
        in: formData
        name: title
        required: true
        type: string
      - description: Description
        in: formData
        name: description
        required: true
        type: string
      - description: Year of release
        in: formData
        name: releaseYear
        required: true
        type: integer
      - description: Director
        in: formData
        name: director
        required: true
        type: string
      - description: Trailer URL
        in: formData
        name: trailerUrl
        required: true
        type: string
      - collectionFormat: csv
        description: Genre ids
        in: formData
        items:
          type: integer
        name: genreIds
        required: true
        type: array
      - description: Poster image
        in: formData
        name: poster
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              id:
                type: integer
            type: object
        "400":
          description: Invalid data
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - Bearer: []
      summary: Update movie
      tags:
      - movies
  /movies/{id}/rate:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Movie id
        in: path
        name: id
        required: true
        type: integer
      - description: Movie rating
        in: query
        name: rating
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid data
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - Bearer: []
      summary: Set movie rating
      tags:
      - movies
  /movies/{id}/setWatched:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Movie id
        in: path
        name: id
        required: true
        type: integer
      - description: Flag value
        in: query
        name: isWatched
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Invalid data
          schema:
            $ref: '#/definitions/models.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ApiError'
      security:
      - Bearer: []
      summary: Mark movie as watched
      tags:
      - movies
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"